+++ packages/html/ds/src/header/header.stories.ts
@@ -21,57 +21,94 @@ const meta = {
 export default meta;
 type Story = StoryObj<typeof meta>;
 
+const slotExample1 = () => `
+  <ul class="gi-list-bullet" data-testid="govieList">
+    <li>
+      <a
+        href="#"
+        class="gi-link gi-link-sm"
+      >
+        Citizens Information - Services and Rights
+      </a>
+    </li>
+    <li>
+      <a
+        href="#"
+        class="gi-link gi-link-sm"
+      >
+        Revenue - Taxes and Payments
+      </a>
+    </li>
+    <li>
+      <a
+        href="#"
+        class="gi-link gi-link-sm"
+      >
+        Department of Social Protection
+      </a>
+    </li>
+  </ul>
+`;
+
+const slotExample2 = () => `
+  <form class="gi-max-w-md gi-mx-auto">
+    <h4 class="gi-heading-sm">Search the website</h4>
+    <div class="gi-flex gi-items-end gi-mt-4">
+      <div class="gi-text-input-container gi-flex-auto">
+        <div class="gi-text-input-container-inner">
+          <input
+            placeholder="Enter search term"
+            id="search"
+            type="text"
+            data-testid="textbox"
+            class="gi-border-gray-950 gi-w-full gi-text-input"
+            name="search_query"
+            aria-label="Search the website"
+          />
+        </div>
+      </div>
+      <button
+        data-testid="govieButton-undefined-undefined-undefined-"
+        class="gi-btn gi-btn-primary gi-btn-regular gi-ml-1 gi-flex-none"
+      >
+        Search
+        <span
+          data-testid="govie-icon"
+          aria-label="Search"
+          role="img"
+          class="material-symbols-outlined gi-block"
+          style="font-size: 24px; font-variation-settings: 'FILL' 0, 'wght' 400, 'GRAD' 0, 'opsz' 24;"
+        >
+          search
+        </span>
+      </button>
+    </div>
+  </form>
+`;
+
+const slotExample3 = () => `
+  <select
+    class="gi-select"
+    id="slot-example-2"
+    aria-label="slot-example-2"
+  >
+    <optgroup label="Languages">
+      <option class="gi-select-option" value="gaeilge">Gaeilge</option>
+      <option class="gi-select-option" value="english">English</option>
+      <option class="gi-select-option" value="spanish">Spanish</option>
+      <option class="gi-select-option" value="italian">Italian</option>
+    </optgroup>
+  </select>
+`;
+
 export const Default: Story = {
-  argTypes: {
-    title: {
-      control: 'text',
-      description: 'The title of the Header',
-      table: {
-        category: 'Header',
-      },
-    },
-    logo: {
-      control: 'object',
-      description: 'The url and image for the logo',
-      table: {
-        category: 'Header',
-      },
-    },
-    tools: {
-      control: 'object',
-      description: 'Actionable items such as Search, Menu and additional CTA',
-      table: {
-        category: 'Header',
-        type: {
-          summary: '{search?: {}; menu?: {}; items: {}[]}',
-        },
-      },
-    },
-    navLinks: {
-      description: 'A list of navigation links',
-      table: {
-        category: 'Header',
-      },
-    },
-    languages: {
-      description: 'A list of secondary navigation links',
-      table: {
-        category: 'Header',
-      },
-    },
-  },
   args: {
-    fullWidth: false,
     logo: {
       href: '/link',
     },
     tools: {
-      search: {
-        action: '/search_page',
-        label: 'Search',
-        icon: 'search',
-      },
       menu: {
+        icon: 'menu',
         label: 'Menu',
       },
       items: [
@@ -79,6 +116,27 @@ export const Default: Story = {
           href: '/item1',
           label: 'Home',
           icon: 'home',
+        },
+        {
+          href: '#',
+          label: 'Faq',
+          icon: 'info',
+          keepOnMobile: true,
+          slot: slotExample1(),
+        },
+        {
+          href: '/search_page',
+          label: 'Search',
+          keepOnMobile: true,
+          icon: 'search',
+          slot: slotExample2(),
+        },
+        {
+          href: '#',
+          label: 'Languages',
+          icon: 'mic',
+          keepOnMobile: true,
+          slot: slotExample3(),
         },
       ],
     },
@@ -90,12 +148,6 @@ export const Default: Story = {
       {
         href: '#',
         label: 'Services',
-      },
-    ],
-    languages: [
-      {
-        href: '#',
-        label: 'Gaeilge',
       },
     ],
   },
@@ -298,7 +350,7 @@ export const tabletView: Story = {
   parameters: {
     layout: 'fullscreen',
     viewport: {
-      defaultViewport: 'pixel',
+      defaultViewport: 'ipad',
     },
   },
   args: {
@@ -491,6 +543,30 @@ export const FullWidth: Story = {
       {
         href: '#',
         label: 'Gaeilge',
+      },
+    ],
+  },
+};
+
+export const ShowMobileMenuForLanguages: Story = {
+  parameters: {
+    layout: 'fullscreen',
+    viewport: {
+      defaultViewport: 'mobile2',
+    },
+  },
+  args: {
+    logo: {
+      href: '/link',
+    },
+    languages: [
+      {
+        href: '#',
+        label: 'Gaeilge',
+      },
+      {
+        href: '#',
+        label: 'English',
       },
     ],
   },