+++ packages/react/ds/CHANGELOG.md
@@ -1,6 +1,42 @@
 # Changelog
 
 
+## [0.32.0](https://github.com/ogcio/govie-ds/compare/react-v0.31.0...react-v0.32.0) (2025-01-07)
+
+
+### Features
+
+* add external link property to header and footer ([#378](https://github.com/ogcio/govie-ds/issues/378)) ([34b726e](https://github.com/ogcio/govie-ds/commit/34b726ea8112ac102f1d985f30559eb2deafcf25))
+* adding slots feature in govieHeader macro ([#373](https://github.com/ogcio/govie-ds/issues/373)) ([95fd203](https://github.com/ogcio/govie-ds/commit/95fd2034aef907084e067628dae18f600b1482b8))
+* composable combo box ([#381](https://github.com/ogcio/govie-ds/issues/381)) ([6d21aa5](https://github.com/ogcio/govie-ds/commit/6d21aa5631903d5faccd400ad27f72e6e78bff20))
+* Progress Bar component ([#369](https://github.com/ogcio/govie-ds/issues/369)) ([83add63](https://github.com/ogcio/govie-ds/commit/83add6397dedf6db84ab18a3f4aa0fc99eb4eebe))
+
+
+### Bug Fixes
+
+* small checkbox alignment ([#383](https://github.com/ogcio/govie-ds/issues/383)) ([7c972a7](https://github.com/ogcio/govie-ds/commit/7c972a76e6c2e5535af579e05f67780d6b48a9fa))
+
+## [0.31.0](https://github.com/ogcio/govie-ds/compare/react-v0.30.0...react-v0.31.0) (2024-12-19)
+
+
+### Features
+
+* select composable ([#376](https://github.com/ogcio/govie-ds/issues/376)) ([8c212b0](https://github.com/ogcio/govie-ds/commit/8c212b02400d66ef8c86c751147275564efefb33))
+
+## [0.30.0](https://github.com/ogcio/govie-ds/compare/react-v0.29.0...react-v0.30.0) (2024-12-18)
+
+
+### Features
+
+* add html and macro for tooltip ([#371](https://github.com/ogcio/govie-ds/issues/371)) ([adebe40](https://github.com/ogcio/govie-ds/commit/adebe405318eb44503817356a429105cc9c0c024))
+
+
+### Bug Fixes
+
+* assets and logo ([#374](https://github.com/ogcio/govie-ds/issues/374)) ([86cf4dc](https://github.com/ogcio/govie-ds/commit/86cf4dcc66aa09bcc252100b0d3fc95687a55021))
+* checkbox feedback fixes ([#372](https://github.com/ogcio/govie-ds/issues/372)) ([842c48d](https://github.com/ogcio/govie-ds/commit/842c48dcf9c75ab45cdcd490371e20387c847c57))
+* show mobile menu for languages  ([#370](https://github.com/ogcio/govie-ds/issues/370)) ([0f258e1](https://github.com/ogcio/govie-ds/commit/0f258e161530e8cce7f2ae9a2e58b2c70fb61e27))
+
 ## [0.29.0](https://github.com/ogcio/govie-ds/compare/react-v0.28.0...react-v0.29.0) (2024-12-12)